+++ CHANGELOG.md
@@ -7,6 +7,7 @@ on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Upcoming Release] - YYYY/MM/DD
 
 - Moved `opynsim.ForwardDynamicSolver` to `opynsim.solvers.ForwardDynamicsSolver`.
+- C++: Added `osc::Camera::Camera(position, direction, up)` constructor.
 
 
 ## 0.0.8 - 2026/09/08